+++ src/config.php
@@ -213,6 +213,9 @@ return [
         //    'putyourlightson\blitz\drivers\integrations\SeomaticIntegration',
         //],
 
+        // Whether to set default no cache headers.
+        //'setDefaultNoCacheHeaders' => true,
+
         // The value to send in the cache control header.
         //'cacheControlHeader' => 'public, s-maxage=31536000, max-age=0',